+++ .metadata.sh
@@ -44,4 +44,4 @@ export BINARY GHUSER HBREPO MAINT VENDOR DESC GOLANGCI_LINT_ARGS CONFIG_FILE
 export LICENSE FORMULA SOURCE_URL URL VERSION_PATH SOURCE_PATH COMPRESS
 
 export WINDOWS_LDFLAGS="-H windowsgui"
-export MACAPP="DiscordNotifier"
+export MACAPP="DiscordNotifier-Client"
