+++ dnclient/tray_other.go
@@ -1,0 +1,5 @@
+// +build !windows,!darwin
+
+package dnclient
+
+func (c *Client) startTray() {}
